adds useAuth custom hooke
